*: remove server/service.go
